requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
torch
spacy
pymupdf
python-multipart
joblib
pytesseract
Pillow
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.25
alembic==1.12.0
pydantic==1.10.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]
bcrypt==3.2.2
email-validator==1.3.1
python-multipart==0.0.6
aiofiles==23.1.0
dotenv